+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineViewTest.kt
@@ -12,7 +12,6 @@ import android.view.View
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import mozilla.components.browser.engine.gecko.GeckoEngineView.Companion.DARK_COVER
 import mozilla.components.browser.engine.gecko.selection.GeckoSelectionActionDelegate
-import mozilla.components.concept.engine.EngineView
 import mozilla.components.concept.engine.mediaquery.PreferredColorScheme
 import mozilla.components.concept.engine.selection.SelectionActionDelegate
 import mozilla.components.support.test.argumentCaptor
@@ -22,6 +21,7 @@ import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertNull
+import org.junit.Assert.assertSame
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -278,55 +278,11 @@ class GeckoEngineViewTest {
         assertFalse(engineView.canClearSelection())
     }
 
-    @Suppress("Deprecation")
     @Test
-    fun `currentInputResult should default to EngineView#InputResult#INPUT_RESULT_UNHANDLED`() {
+    fun `GIVEN a GeckoView WHEN EngineView returns the InputResultDetail THEN the value from the GeckoView is used`() {
         val engineView = GeckoEngineView(context)
+        val geckoview = engineView.geckoView
 
-        assertEquals(EngineView.InputResult.INPUT_RESULT_UNHANDLED, engineView.lastInputResult)
-    }
-
-    @Suppress("Deprecation")
-    @Test
-    fun `getInputResult should do a 1-1 mapping of the values received from GeckoView and cache the result`() {
-        val engineView = GeckoEngineView(context)
-        engineView.geckoView = mock()
-
-        whenever(engineView.geckoView.inputResult).thenReturn(0)
-        assertEquals(EngineView.InputResult.INPUT_RESULT_UNHANDLED, engineView.getInputResult())
-        assertEquals(EngineView.InputResult.INPUT_RESULT_UNHANDLED, engineView.lastInputResult)
-
-        whenever(engineView.geckoView.inputResult).thenReturn(1)
-        assertEquals(EngineView.InputResult.INPUT_RESULT_HANDLED, engineView.getInputResult())
-        assertEquals(EngineView.InputResult.INPUT_RESULT_HANDLED, engineView.lastInputResult)
-
-        whenever(engineView.geckoView.inputResult).thenReturn(2)
-        assertEquals(EngineView.InputResult.INPUT_RESULT_HANDLED_CONTENT, engineView.getInputResult())
-        assertEquals(EngineView.InputResult.INPUT_RESULT_HANDLED_CONTENT, engineView.lastInputResult)
-    }
-
-    @Suppress("Deprecation")
-    @Test
-    fun `INPUT_RESULD_IGNORED should be ignored`() {
-        val engineView = GeckoEngineView(context)
-        engineView.geckoView = mock()
-
-        whenever(engineView.geckoView.inputResult).thenReturn(1)
-        assertEquals(EngineView.InputResult.INPUT_RESULT_HANDLED, engineView.getInputResult())
-        assertEquals(EngineView.InputResult.INPUT_RESULT_HANDLED, engineView.lastInputResult)
-
-        whenever(engineView.geckoView.inputResult).thenReturn(3)
-        assertEquals(EngineView.InputResult.INPUT_RESULT_HANDLED, engineView.getInputResult())
-        assertEquals(EngineView.InputResult.INPUT_RESULT_HANDLED, engineView.lastInputResult)
-    }
-
-    @Suppress("Deprecation")
-    @Test(expected = IllegalArgumentException::class)
-    fun `Values other than 0, 1, 2, 3 received as input results from GeckoView should throw`() {
-        val engineView = GeckoEngineView(context)
-        engineView.geckoView = mock()
-
-        whenever(engineView.geckoView.inputResult).thenReturn(4)
-        engineView.getInputResult()
+        assertSame(geckoview.inputResultDetail, engineView.getInputResultDetail())
     }
 }