+++ mobile/android/android-components/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/NestedWebView.kt
@@ -6,18 +6,20 @@ package mozilla.components.browser.engine.system
 
 import android.annotation.SuppressLint
 import android.content.Context
-import androidx.core.view.NestedScrollingChild
-import androidx.core.view.NestedScrollingChildHelper
 import android.view.MotionEvent
-import android.view.MotionEvent.ACTION_MOVE
-import android.view.MotionEvent.ACTION_DOWN
-import android.view.MotionEvent.ACTION_UP
 import android.view.MotionEvent.ACTION_CANCEL
+import android.view.MotionEvent.ACTION_DOWN
+import android.view.MotionEvent.ACTION_MOVE
+import android.view.MotionEvent.ACTION_UP
 import android.view.MotionEvent.obtain
 import android.webkit.WebView
 import androidx.annotation.VisibleForTesting
+import androidx.core.view.NestedScrollingChild
+import androidx.core.view.NestedScrollingChildHelper
 import androidx.core.view.ViewCompat
-import mozilla.components.concept.engine.EngineView
+import mozilla.components.concept.engine.INPUT_HANDLED
+import mozilla.components.concept.engine.INPUT_UNHANDLED
+import mozilla.components.concept.engine.InputResultDetail
 
 /**
  * WebView that supports nested scrolls (for using in a CoordinatorLayout).
@@ -46,18 +48,12 @@ class NestedWebView(context: Context) : WebView(context), NestedScrollingChild {
     internal var childHelper: NestedScrollingChildHelper = NestedScrollingChildHelper(this)
 
     /**
-     * Integer indicating how user's MotionEvent was handled.
+     * How user's MotionEvent will be handled.
      *
-     * There must be a 1-1 relation between this values and [EngineView.InputResult]'s.
+     * @see InputResultDetail
      */
-    @Suppress("Deprecation")
-    internal val inputResult: Int
-        get() = getInputResult(eventHandled)
-
-    /**
-     * Holder for if the previous [android.view.MotionEvent] was handled by us or not.
-     */
-    private var eventHandled: Boolean = false
+    @VisibleForTesting
+    internal var inputResultDetail = InputResultDetail.newInstance()
 
     init {
         isNestedScrollingEnabled = true
@@ -104,12 +100,18 @@ class NestedWebView(context: Context) : WebView(context), NestedScrollingChild {
         }
 
         // Execute event handler from parent class in all cases
-        eventHandled = super.onTouchEvent(event)
+        val eventHandled = callSuperOnTouchEvent(event)
+        updateInputResult(eventHandled)
 
         // Recycle previously obtained event
         event.recycle()
 
         return eventHandled
+    }
+
+    @VisibleForTesting
+    internal fun callSuperOnTouchEvent(event: MotionEvent): Boolean {
+        return super.onTouchEvent(event)
     }
 
     // NestedScrollingChild
@@ -156,12 +158,14 @@ class NestedWebView(context: Context) : WebView(context), NestedScrollingChild {
         return childHelper.dispatchNestedPreFling(velocityX, velocityY)
     }
 
-    @Suppress("Deprecation")
-    private fun getInputResult(eventHandled: Boolean): Int {
-        return if (eventHandled) {
-            EngineView.InputResult.INPUT_RESULT_HANDLED.value
-        } else {
-            EngineView.InputResult.INPUT_RESULT_UNHANDLED.value
-        }
+    @VisibleForTesting
+    internal fun updateInputResult(eventHandled: Boolean) {
+        inputResultDetail = inputResultDetail.copy(
+            if (eventHandled) {
+                INPUT_HANDLED
+            } else {
+                INPUT_UNHANDLED
+            }
+        )
     }
 }