+++ mobile/android/android-components/components/browser/toolbar/src/test/java/mozilla/components/browser/toolbar/behavior/BrowserToolbarBehaviorTest.kt
@@ -17,6 +17,7 @@ import com.google.android.material.snackbar.Snackbar
 import mozilla.components.browser.toolbar.BrowserToolbar
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.EngineView
+import mozilla.components.concept.engine.InputResultDetail
 import mozilla.components.concept.engine.selection.SelectionActionDelegate
 import mozilla.components.support.test.any
 import mozilla.components.support.test.mock
@@ -100,6 +101,29 @@ class BrowserToolbarBehaviorTest {
             type = ViewCompat.TYPE_TOUCH,
             toolbar = mock()
         )
+        assertFalse(acceptsNestedScroll)
+    }
+
+    @Test
+    fun `GIVEN a gesture that doesn't scroll the toolbar WHEN startNestedScroll THEN toolbar is expanded and nested scroll not accepted`() {
+        val behavior = spy(BrowserToolbarBehavior(testContext, null, ToolbarPosition.BOTTOM))
+        val engineView: EngineView = mock()
+        val inputResultDetail: InputResultDetail = mock()
+        val yTranslator: BrowserToolbarYTranslator = mock()
+        behavior.yTranslator = yTranslator
+        doReturn(false).`when`(behavior).shouldScroll
+        doReturn(true).`when`(inputResultDetail).isTouchUnhandled()
+        behavior.engineView = engineView
+        doReturn(inputResultDetail).`when`(engineView).getInputResultDetail()
+
+        val acceptsNestedScroll = behavior.startNestedScroll(
+            axes = ViewCompat.SCROLL_AXIS_VERTICAL,
+            type = ViewCompat.TYPE_TOUCH,
+            toolbar = mock()
+        )
+
+        verify(yTranslator).cancelInProgressTranslation()
+        verify(yTranslator).expandWithAnimation(any())
         assertFalse(acceptsNestedScroll)
     }
 
@@ -302,74 +326,71 @@ class BrowserToolbarBehaviorTest {
         verify(behavior).tryToScrollVertically(-100f)
     }
 
-    @Suppress("Deprecation")
     @Test
-    fun `Behaviour shouldScroll if EngineView handled the MotionEvent and scrolling is enabled`() {
+    fun `GIVEN a null InputResultDetail from the EngineView WHEN shouldScroll is called THEN it returns false`() {
+        val behavior = BrowserToolbarBehavior(testContext, null, ToolbarPosition.BOTTOM)
+        behavior.engineView = null
+        assertFalse(behavior.shouldScroll)
+        behavior.engineView = mock()
+        `when`(behavior.engineView!!.getInputResultDetail()).thenReturn(null)
+
+        assertFalse(behavior.shouldScroll)
+    }
+
+    @Test
+    fun `GIVEN an InputResultDetail with the right values and scroll enabled WHEN shouldScroll is called THEN it returns true`() {
         val behavior = BrowserToolbarBehavior(testContext, null, ToolbarPosition.BOTTOM)
         val engineView: EngineView = mock()
-        `when`(engineView.getInputResult()).thenReturn(EngineView.InputResult.INPUT_RESULT_HANDLED)
         behavior.engineView = engineView
         behavior.isScrollEnabled = true
+        val validInputResultDetail: InputResultDetail = mock()
+        doReturn(validInputResultDetail).`when`(engineView).getInputResultDetail()
 
+        doReturn(true).`when`(validInputResultDetail).canScrollToBottom()
+        doReturn(false).`when`(validInputResultDetail).canScrollToTop()
+        assertTrue(behavior.shouldScroll)
+
+        doReturn(false).`when`(validInputResultDetail).canScrollToBottom()
+        doReturn(true).`when`(validInputResultDetail).canScrollToTop()
+        assertTrue(behavior.shouldScroll)
+
+        doReturn(true).`when`(validInputResultDetail).canScrollToBottom()
+        doReturn(true).`when`(validInputResultDetail).canScrollToTop()
         assertTrue(behavior.shouldScroll)
     }
 
-    @Suppress("Deprecation")
     @Test
-    fun `Behaviour !shouldScroll if EngineView handled the MotionEvent and scrolling is !enabled`() {
+    fun `GIVEN an InputResultDetail with the right values but with scroll disabled WHEN shouldScroll is called THEN it returns false`() {
         val behavior = BrowserToolbarBehavior(testContext, null, ToolbarPosition.BOTTOM)
-        val engineView: EngineView = mock()
-        `when`(engineView.getInputResult()).thenReturn(EngineView.InputResult.INPUT_RESULT_HANDLED)
-        behavior.engineView = engineView
+        behavior.engineView = mock()
         behavior.isScrollEnabled = false
+        val validInputResultDetail: InputResultDetail = mock()
+        doReturn(true).`when`(validInputResultDetail).canScrollToBottom()
+        doReturn(true).`when`(validInputResultDetail).canScrollToTop()
 
         assertFalse(behavior.shouldScroll)
     }
 
-    @Suppress("Deprecation")
     @Test
-    fun `Behaviour !shouldScroll if EngineView didn't handle the MotionEvent and scrolling is enabled`() {
+    fun `GIVEN scroll enabled but EngineView cannot scroll to bottom WHEN shouldScroll is called THEN it returns false`() {
         val behavior = BrowserToolbarBehavior(testContext, null, ToolbarPosition.BOTTOM)
-        val engineView: EngineView = mock()
-        `when`(engineView.getInputResult()).thenReturn(EngineView.InputResult.INPUT_RESULT_UNHANDLED)
-        behavior.engineView = engineView
+        behavior.engineView = mock()
         behavior.isScrollEnabled = true
+        val validInputResultDetail: InputResultDetail = mock()
+        doReturn(false).`when`(validInputResultDetail).canScrollToBottom()
+        doReturn(true).`when`(validInputResultDetail).canScrollToTop()
 
         assertFalse(behavior.shouldScroll)
     }
 
-    @Suppress("Deprecation")
     @Test
-    fun `Behaviour !shouldScroll if EngineView didn't handle the MotionEvent and scrolling is !enabled`() {
+    fun `GIVEN scroll enabled but EngineView cannot scroll to top WHEN shouldScroll is called THEN it returns false`() {
         val behavior = BrowserToolbarBehavior(testContext, null, ToolbarPosition.BOTTOM)
-        val engineView: EngineView = mock()
-        `when`(engineView.getInputResult()).thenReturn(EngineView.InputResult.INPUT_RESULT_UNHANDLED)
-        behavior.engineView = engineView
-        behavior.isScrollEnabled = false
-
-        assertFalse(behavior.shouldScroll)
-    }
-
-    @Suppress("Deprecation")
-    @Test
-    fun `Behaviour !shouldScroll if EngineView didn't handle the MotionEvent but the website and scrolling is enabled`() {
-        val behavior = BrowserToolbarBehavior(testContext, null, ToolbarPosition.BOTTOM)
-        val engineView: EngineView = mock()
-        `when`(engineView.getInputResult()).thenReturn(EngineView.InputResult.INPUT_RESULT_HANDLED_CONTENT)
-        behavior.engineView = engineView
+        behavior.engineView = mock()
         behavior.isScrollEnabled = true
-
-        assertFalse(behavior.shouldScroll)
-    }
-
-    @Suppress("Deprecation")
-    @Test
-    fun `Behaviour !shouldScroll if EngineView didn't handle the MotionEvent but the website and scrolling is !enabled`() {
-        val behavior = BrowserToolbarBehavior(testContext, null, ToolbarPosition.BOTTOM)
-        val engineView: EngineView = mock()
-        `when`(engineView.getInputResult()).thenReturn(EngineView.InputResult.INPUT_RESULT_HANDLED_CONTENT)
-        behavior.engineView = engineView
-        behavior.isScrollEnabled = false
+        val validInputResultDetail: InputResultDetail = mock()
+        doReturn(true).`when`(validInputResultDetail).canScrollToBottom()
+        doReturn(false).`when`(validInputResultDetail).canScrollToTop()
 
         assertFalse(behavior.shouldScroll)
     }